requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
sentence-transformers>=2.2
hdbscan>=0.8
umap-learn>=0.5
matplotlib>=3.7
seaborn>=0.12
wordcloud>=1.9
pyyaml>=6.0
scipy>=1.11
pytest>=7.4
jupyterlab>=4.0
nbformat>=5.9
